extls: 0.4.6-alpha => 0.5-beta
Markup: added Line(...); split into partial Markup.cs, Markup.Line.cs, Markup.Utils.cs, and other improvements;
Arguments: added a central argument parser for modules;
Arguments.Args: added the Args class for Arguments;
Print: renamed to Out;
Deleted all Tools.* modules except Dir;
ModuleShell: renamed from ModuleMenu;
ModuleRaw: added DispatchRaw2(...);
Program: upgraded to the new architecture;
Dir: upgraded to the new architecture;
Color: added new base colors and other improvements;

New:
- ModuleMenu: new terminal UI framework, currently in development;
- Gradient;
- Root: global base module; use `root` for extls root commands;

+++ extls/Tools/Dir/Dir.cs
@@ -21,30 +21,19 @@ public record DirConfig(bool icons, bool mini, string[] ignoreExtensions)
 [ModuleName("dir")]
 public partial class Dir : Module
 {
-    private DirConfig config = new(false, false, [".testextension"]);
+    private DirConfig config = new(false, false, [".ignoreextension"]);
 
     public Dir()
     {
         name = "dir";
-        version = "0.6.8b";
-        commands =
-        [
-            new HelpSlot("tree", "scans the specified directory",
-                      new[] {"\"--recursive, -r\" - recursive path scanning",
-                             "\"--summary, --sum, -s\" - scan quantity only",
-                             "\"--type file\\folder\\all\" - specify scanning of folders, files, or both (default: all)"},
-                             "\"extls dir tree \"YOUR_PATH_ON_DISK\" -s --type folder\""),
-                new HelpSlot("create", "creates files\\folders on the disk at the specified path or relative to the terminal's current path",
-                      new[] {"\"--set-path, -sp\" - manually specify the full path (e.g., \"-sp Disk:\\\")"},
-                             "\'extls dir create file\\\\folder \"file.txt\" -sp \"PATH\"\'")
-        ];
+        version = "0.6.9b";
 
-        string path = Path.Combine(Root.RootPath, "config", "dir-config.json");
+        string path = Path.Combine(Global.RootPath, "config", "dir-config.json");
 
         if (File.Exists(path))
         {
             config = JsonService.LoadJson<DirConfig>(
-                Path.Combine(Root.RootPath, "config"),
+                Path.Combine(Global.RootPath, "config"),
                 "dir-config.json");
         }
         else
@@ -58,7 +47,7 @@ public partial class Dir : Module
             }
 
             JsonService.SaveJson(
-                Path.Combine(Root.RootPath, "config"),
+                Path.Combine(Global.RootPath, "config"),
                 "dir-config.json",
                 config);
 
@@ -67,14 +56,14 @@ public partial class Dir : Module
         }
     }
 
-    [MethodName(Params.Args, ["create", "c"])]
+    [MethodName(["create", "c"])]
     public void Create(string[] args)
     {
-        if (args.Length == 0) { Print.Error("Arguments are missing."); return; }
+        if (args.Length == 0) { Out.Error("Arguments are missing."); return; }
 
         int createType = args[0] switch { "file" => 1, "folder" => 0, _ => -1, };
-        if (createType == -1) { Print.Error("Invalid type for creation."); return; }
-        if (args.Length < 2) { Print.Error("Invalid name for creation."); }
+        if (createType == -1) { Out.Error("Invalid type for creation."); return; }
+        if (args.Length < 2) { Out.Error("Invalid name for creation."); }
         string name = args[1];
 
         bool auto = true;
@@ -93,7 +82,7 @@ public partial class Dir : Module
         }
 
         string path = auto ? Directory.GetCurrentDirectory() : userPath;
-        if (!Path.Exists(path)) { Print.Error("Unknown path."); return; }
+        if (!Path.Exists(path)) { Out.Error("Unknown path."); return; }
 
         bool success = true;
         string failReason = "";
@@ -115,93 +104,65 @@ public partial class Dir : Module
         catch (Exception ex) { success = false; failReason = Reason(ex); }
 
         if (success) Markup.Rich($"Item created **$[green]successfully[white]**!" +
-                                 $"\n$[white]Item name: $[cyan]{Markup.FixBackslash(name)}" +
-                                 $"\n$[white]Path: $[yellow]{Markup.FixBackslash(path)}");
+                                 $"\n$[white]Item name: $[cyan]{Markup.SafeBackslash(name)}" +
+                                 $"\n$[white]Path: $[yellow]{Markup.SafeBackslash(path)}");
         else Markup.Rich($"Create item **$[red]failed$[white]** of reason: $[yellow]{failReason}$[yellow]." +
-                         $"\n$[white]Item name: **$[cyan]{Markup.FixBackslash(name)}**" +
-                         $"\n$[white]Path: *$[yellow]{Markup.FixBackslash(path)}*");
+                         $"\n$[white]Item name: **$[cyan]{Markup.SafeBackslash(name)}**" +
+                         $"\n$[white]Path: *$[yellow]{Markup.SafeBackslash(path)}*");
     }
 
-    [MethodName(Params.Args, ["tree", "tr"])]
-    public void Tree(string[] args)
+    [MethodName(["tree", "tr"])]
+    public void Tree()
     {
-        if (args.Length == 0)
+        bool recursive = Arguments.Get("r", "recursive");
+
+        if (int.TryParse(Arguments.GetRight("r", "recursive"), out int recursiveLevels)){}
+        else recursiveLevels = recursive ? 999999 : 1;
+        
+        bool summary = Arguments.Get("s", "sum", "summary");
+        byte typeFilter = Arguments.GetRight("t", "type") switch {"file" => 2, "folder" => 1, _ or "all" => 0};
+
+        string? path = Arguments.GetPath();
+
+        if (path is null)
         {
-            Print.Error("Arguments are missing.");
-            return;
-        }
-
-        bool recursive = false;
-        int recursiveLevels = 1;
-        bool summary = false;
-        byte foldersOrFilesOrAll = 0; // 0 - all, 1 - folder, 2 - file
-
-        for (int i = 0; i < args.Length; i++)
-        {
-            switch (args[i])
-            {
-                case "--recursive" or "-r":
-                    recursive = true;
-                    recursiveLevels = 999999;
-                    if (i + 1 < args.Length && int.TryParse(args[i + 1], out int l))
-                    {
-                        recursiveLevels = l;
-                        continue;
-                    }
-                    break;
-                case "--summary" or "--sum" or "-s": summary = true; break;
-                case "--type":
-                    if (i + 1 < args.Length)
-                        foldersOrFilesOrAll = args[i + 1] switch
-                        {
-                            "file" => 2,
-                            "folder" => 1,
-                            "all" => 0,
-                            _ => 0
-                        };
-                    break;
-            }
-        }
-
-        string path = args[0];
-
-        if (args[0] is ".") path = Directory.GetCurrentDirectory();
-        else if (args.Length > 0 && args[0][0] is '.')
-            path = Path.GetFullPath(args[0], Directory.GetCurrentDirectory());
-
-        if (!Directory.Exists(path))
-        {
-            Print.Error("Unknown path.");
-            Print.Debug(path);
+            Out.Error("Unknown path.");
             return;
         }
 
         if (!config!.mini)
             Markup.Rich($"Scan result of path $[{
-                (Root.Platform is Platform.Windows ? "yellow" : "#7dbeff")
-            }]'{Markup.FixBackslash(path)}'$[white]:\n", true);
+                (Global.Platform is Platform.Windows ? "yellow" : "#7dbeff")
+            }]'{Markup.SafeBackslash(path)}'$[white]:\n", true);
         else Console.WriteLine();
 
         Stopwatch time = Stopwatch.StartNew();
 
-        int sumDir = 0;
-        int sumF = 0;
+        int summaryFolders = 0;
+        int summaryFiles = 0;
 
-        ExecuteTree(path, 0, (recursive ? recursiveLevels + 1 : 1), summary, foldersOrFilesOrAll, ref sumDir, ref sumF);
+        ExecuteTree(
+            path: path,
+            currentLevel: 0,
+            maxLevels: recursive ? recursiveLevels + 1 : 1,
+            summary: summary,
+            typeFilter: typeFilter,
+            summaryFolders: ref summaryFolders,
+            summaryFiles: ref summaryFiles
+        );
 
-        if (summary || Print.verbose)
+        if (summary || Global.Verbose)
         {
-            if (foldersOrFilesOrAll is 0 or 1) Markup.Rich($"Folders $[green]scanned$[white]: {sumDir}\n");
-            if (foldersOrFilesOrAll is 0 or 2) Markup.Rich($"Files $[green]scanned$[white]: {sumF}\n");
+            if (typeFilter is 0 or 1) Markup.Rich($"Folders $[green]scanned$[white]: {summaryFolders}\n");
+            if (typeFilter is 0 or 2) Markup.Rich($"Files $[green]scanned$[white]: {summaryFiles}\n");
         }
 
         time.Stop();
-        if (Print.verbose)
-            Markup.Rich($"\nScanned in: $[#d375ff]{time.Elapsed}", true);
+        if (Global.Verbose) Markup.Rich($"\nScanned in: $[#d375ff]{time.Elapsed}", true);
         Console.WriteLine();
     }
 
-    [MethodName(Params.None, ["grid", "gr", "grd"])]
+    [MethodName(["grid", "gr", "grd"])]
     public void Grid()
     {
         string path = Directory.GetCurrentDirectory();
@@ -218,7 +179,7 @@ public partial class Dir : Module
 
         if (items.Count == 0)
         {
-            Print.Warning("The current folder is empty.");
+            Out.Warning("The current folder is empty.");
             return;
         }
 
@@ -242,8 +203,8 @@ public partial class Dir : Module
         });
 
         foreach (GridItem item in items)
-            Print.Debug($"{item.name} ({item.name.Length} chars.)");
-        Print.Debug("\n");
+            Out.Debug($"{item.name} ({item.name.Length} chars.)");
+        Out.Debug("\n");
 
         int width = Console.WindowWidth;
         int lengthCount = 0;
@@ -263,30 +224,18 @@ public partial class Dir : Module
 
             lengthCount += output.Length;
 
-            Markup.Rich($"$[{(items[i].isFolder ? (Root.Platform is Platform.Linux ? "#7dbeff" : "yellow") : fileIconUsable.color)}]{output}");
+            Markup.Rich($"$[{(items[i].isFolder ? (Global.Platform is Platform.Linux ? "#7dbeff" : "yellow") : fileIconUsable.color)}]{output}");
         }
     }
 
 
-    [MethodName(Params.Args, "config")]
-    public void Config(string[] args)
+    [MethodName("config")]
+    public void Config()
     {
-        string path = Path.Combine(Root.RootPath, "config", "dir-config.json");
+        string path = Path.Combine(Global.RootPath, "config", "dir-config.json");
 
-        if (args.Length > 0)
-        {
-            for (int i = 0; i < args.Length; i++)
-            {
-                switch (args[i])
-                {
-                    case "--clear" or "-c":
-                        if (File.Exists(path))
-                            File.Delete(path);
-                        break;
-                }
-            }
-        }
-
+        if (Arguments.Get("c", "clear"))
+            if (File.Exists(path)) File.Delete(path);        
 
         bool ingoreExtensions = config.ignoreExtensions.Length > 0;
         string extensions = string.Empty;
